Add gallery DB configuration and DI to ExtractAndValidateSignature
