+++ src/NuGet.Jobs.Common/Configuration/GalleryDbConfiguration.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace NuGet.Services.Validation.Orchestrator
+namespace NuGet.Jobs.Configuration
 {
     public class GalleryDbConfiguration
     {
