+++ src/Validation.PackageSigning.ExtractAndValidateSignature/Job.cs
@@ -23,6 +23,7 @@ using NuGet.Services.KeyVault;
 using NuGet.Services.ServiceBus;
 using NuGet.Services.Validation;
 using NuGet.Services.Validation.PackageSigning;
+using NuGetGallery;
 
 namespace NuGet.Jobs.Validation.PackageSigning.ExtractAndValidateSignature
 {
@@ -38,6 +39,7 @@ namespace NuGet.Jobs.Validation.PackageSigning.ExtractAndValidateSignature
         /// </summary>
         private const string ConfigurationArgument = "Configuration";
 
+        private const string GalleryDbConfigurationSectionName = "GalleryDb";
         private const string ValidationDbConfigurationSectionName = "ValidationDb";
         private const string ServiceBusConfigurationSectionName = "ServiceBus";
         private const string PackageDownloadTimeoutName = "PackageDownloadTimeout";
@@ -133,6 +135,7 @@ namespace NuGet.Jobs.Validation.PackageSigning.ExtractAndValidateSignature
 
         private void ConfigureJobServices(IServiceCollection services, IConfigurationRoot configurationRoot)
         {
+            services.Configure<GalleryDbConfiguration>(configurationRoot.GetSection(GalleryDbConfigurationSectionName));
             services.Configure<ValidationDbConfiguration>(configurationRoot.GetSection(ValidationDbConfigurationSectionName));
             services.Configure<ServiceBusConfiguration>(configurationRoot.GetSection(ServiceBusConfigurationSectionName));
 
@@ -143,6 +146,15 @@ namespace NuGet.Jobs.Validation.PackageSigning.ExtractAndValidateSignature
                 var config = p.GetRequiredService<IOptionsSnapshot<ValidationDbConfiguration>>().Value;
 
                 return new ValidationEntitiesContext(config.ConnectionString);
+            });
+
+            services.AddTransient<IEntityRepository<Certificate>, EntityRepository<Certificate>>();
+
+            services.AddScoped<IEntitiesContext>(p =>
+            {
+                var config = p.GetRequiredService<IOptionsSnapshot<GalleryDbConfiguration>>().Value;
+
+                return new EntitiesContext(config.ConnectionString, readOnly: true);
             });
 
             services.AddTransient<ISubscriptionClient>(p =>
